tests/test-timing-functions.html

<!-- 
Copyright 2013 Google Inc. All Rights Reserved.

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.

Written by Steph McArthur

This will check if the timing functions create the desired acceleration over the time period
-->

<!DOCTYPE html>
<style>
.anim {
  left: 0px;
  width: 200px;
  height: 24px;
  background-color: lightsteelblue;
  position: relative;
  border-bottom: solid white 1px;
}

#expectation {
  position: absolute;
  left: 200px;
  width: 200px;
  height: 200px;
  background: red;
}
body { margin: 0px; }
</style>

<head>
  <title>Test Timing Functions</title>
  <link rel="author" title="Steph McArthur", href="mailto:stephm@google.com">
  <link rel="help" href="https://dvcs.w3.org/hg/FXTF/raw-file/tip/web-anim/index.html#the-timingfunction-interface">
  <meta name="flags" content="dom">
  <meta name="assert" content="Verify the objects move at the correct speed(s)"
</head>

<body>
  <p>*insert heading here*</p>
</body>

<div id=expectation> </div>

<div> 
  <div class="anim test">ease-in</div>
  <div class="anim test">ease-out</div>
  <div class="anim test">cubic-bezier(0, 0.2, 0.8, 1)</div>
  <div class="anim test">cubic-bezier(0.5, -1.5, 0.5, 2.5)</div>

</div>

<script src="../../web-animations-js/web-animation.js"></script>
<script src="../testharness/testharness.js"></script>
<script src='../testharness/testharnessreport.js'></script>
<script src='../extra-asserts.js'></script>
<link rel='stylesheet' href='../testharness/testharness.css'>
<link rel='stylesheet' type="text/css" href="../animation-test-style.css">

<script>
var elems = document.querySelectorAll(".test");
var anims = [];
for (var i = 0; i < elems.length; i++) {
  var element = elems[i];
  anims[i] = new Animation(element, {left: "200px"}, {startDelay: 1, duration: 2, timingFunction: element.textContent});
}

var endTest = async_test("Testing final locations");
setTimeout(function (){
  endTest.step(function(){
    var eels = document.querySelectorAll(".test");
    for(var i = 0; i < eels.length; i++){
      var params = document.defaultView.getComputedStyle(eels[i], null);
      assert_equals(params.left, "200px", "Cube " + eels[i].textContent +" didn't finish at the right location");
    }
  });
  endTest.done();
  setLog(anims[anims.length-1])
}, 3500);

var stepCount = async_test("Checking the animation has the right number of steps");
var numAnim = (document.querySelectorAll(".test")).length;
var steps = [];
var prevLeft = [];
for(var i = 0; i < numAnim; i++){
  steps[i] = 0;
  prevLeft[i] = 0;
}
var counter = setInterval(checkForStep, 10);

function checkForStep (){
  var eels = document.querySelectorAll(".test");
  for(var i = 0; i < eels.length; i++){
    var params = document.defaultView.getComputedStyle(eels[i], null);
    if(params.left != prevLeft[i]) steps[i]++;
    prevLeft[i] = params.left;
  }
}

setTimeout(function(){
  stepCount.step(function(){
    var eels = document.querySelectorAll(".test");
    for(var i = 0; i < eels.length; i++){
      var params = document.defaultView.getComputedStyle(eels[i], null);
      var numSteps = eels[i].textContent;
      numSteps = numSteps.replace(/[^0-9]/g, " ");
      assert_equals(steps[i] -1, parseInt(numSteps),  eels[i].textContent +" didn't have the right steps");
    }
  });
  stepCount.done();
}, 3500);


//====================================\\
// 13thParallel.org Beziér Curve Code \\
//   by Dan Pupius (www.pupius.net)   \\
//====================================\\

coord = function (x,y) {
  if(!x) var x=0;
  if(!y) var y=0;
  return {x: x, y: y};
}

function B1(t) { return t*t*t }
function B2(t) { return 3*t*t*(1-t) }
function B3(t) { return 3*t*(1-t)*(1-t) }
function B4(t) { return (1-t)*(1-t)*(1-t) }

function getBezier(percent,C1,C2,C3,C4) {
  var pos = new coord();
  pos.x = C1.x*B1(percent) + C2.x*B2(percent) + C3.x*B3(percent) + C4.x*B4(percent);
  pos.y = C1.y*B1(percent) + C2.y*B2(percent) + C3.y*B3(percent) + C4.y*B4(percent);
  return pos;
}


</script>

<div id="log"></div>
